Support graphql-ruby v1.12
