+++ .rubocop.yml
@@ -1,2 +1,5 @@
 inherit_from:
-  - https://onehq.com/rubocop.yml
+  - https://onehq.com/rubocop3.yml
+
+AllCops:
+  SuggestExtensions: false
