+++ README.md
@@ -8,6 +8,7 @@ OneHQ GraphQL interface to [Ruby Graphql](https://github.com/rmosolgo/graphql-ru
 ## Configuration
 
 ### Default Scope
+
 Define a global default scope.
 
 ```ruby
@@ -19,6 +20,7 @@ end
 ```
 
 ### Excluded Inputs
+
 Define a global excluded input fields. Useful for excluding (autogenerated | auto set) fields like below.
 
 ```ruby
@@ -28,9 +30,11 @@ end
 ```
 
 ## GraphQL Resource
+
 Connect to ActiveRecord to auto generate queries and mutations.
 
 ### Queries
+
 Include `::HQ::GraphQL::Resource` and set `self.model_name` to start using queries.
 Fields will be created for all active record columns. Association fields will be created if the association
 is also a GraphQL Resource.
@@ -45,6 +49,7 @@ end
 ```
 
 #### Customize
+
 ```ruby
 class AdvisorResource
   include ::HQ::GraphQL::Resource
@@ -66,6 +71,7 @@ end
 ```
 
 ### Mutations
+
 Mutations will not be created by default. Add `mutations` to a resource to build mutations for create, update, and destroy.
 
 ```ruby
@@ -88,34 +94,40 @@ end
 ```
 
 ### Enums
+
 Auto generate enums from the database using ActiveRecord
 This comes in handy when we have constants that we want represented as enums.
 
 #### Example
+
 Let's assume we're saving data into a user types table
+
 ```postgresl
 # select * from user_types;
-  id |    name    
+  id |    name
  --- +-------------
-  1  | Admin  
+  1  | Admin
   2  | Support User
  (2 rows)
 ```
 
 ```ruby
-class Enums::UserType < ::HQ::GraphQL::Enum
+class Enums::UserType < ::GraphQL::Schema::Enum
   with_model
 end
 ```
+
 This class automatically uses the UserType ActiveRecord model to generate an enum:
+
 ```graphql
-  enum UserType {
-    Admin
-    SupportUser
-  }
+enum UserType {
+  Admin
+  SupportUser
+}
 ```
 
 ### Root Mutations
+
 Add mutations to your schema
 
 ```ruby
@@ -127,7 +139,9 @@ end
 ```
 
 ### Default Root Queries
+
 Create a root query:
+
 ```ruby
 class AdvisorResource
   include ::HQ::GraphQL::Resource
@@ -144,6 +158,7 @@ end
 ```
 
 ### Custom Root Queries
+
 ```ruby
 class AdvisorResource
   include ::HQ::GraphQL::Resource
@@ -163,9 +178,10 @@ class AdvisorResource
 end
 ```
 
-## Create a new ::HQ::GraphQL::Object
+## Create a new ::GraphQL::Schema::Object
+
 ```ruby
-class AdvisorType < ::HQ::GraphQL::Object
+class AdvisorType < ::GraphQL::Schema::Object
   # Supports graphql-ruby functionality
   field :id, Int, null: false
