+++ hq-graphql.gemspec
@@ -14,17 +14,16 @@ Gem::Specification.new do |s|
   s.summary     = "OneHQ GraphQL Library"
   s.description = "OneHQ GraphQL Library"
   s.license     = "MIT"
-  s.required_ruby_version = "~> 2.6"
+  s.required_ruby_version = ">= 2.6"
 
   s.files = Dir["{lib}/**/*", "MIT-LICENSE", "Rakefile", "README.md"]
 
   s.add_dependency "rails",                     "~> 6.0"
 
-  s.add_dependency "graphql",                   "~> 1.9.6"
+  s.add_dependency "graphql",                   "~> 1.12"
   s.add_dependency "graphql-batch",             "~> 0.4"
-  s.add_dependency "graphql-schema_comparator", "~> 0.6"
+  s.add_dependency "graphql-schema_comparator", "~> 1.0"
   s.add_dependency "pg",                        "~> 1.1"
-  s.add_dependency "sprockets",                 "< 4.0.0"
 
   s.add_development_dependency "byebug",                                  "~> 11.0"
   s.add_development_dependency "combustion",                              "~> 1.1"
@@ -32,8 +31,8 @@ Gem::Specification.new do |s|
   s.add_development_dependency "factory_bot_rails",                       "~> 5.0"
   s.add_development_dependency "faker",                                   "~> 2.1"
   s.add_development_dependency "rspec",                                   "~> 3.8"
-  s.add_development_dependency "rspec-rails",                             ">= 4.0.0beta2"
-  s.add_development_dependency "rubocop",                                 "~> 0.7"
+  s.add_development_dependency "rspec-rails",                             "~> 4.0"
+  s.add_development_dependency "rubocop",                                 "~> 1.8"
   s.add_development_dependency "rubocop-performance",                     "~> 1.4"
   s.add_development_dependency "rubocop-rails",                           "~> 2.3"
 end
