+++ lib/hq/graphql/association_loader.rb
@@ -4,6 +4,7 @@ module HQ
   module GraphQL
     class AssociationLoader < ::GraphQL::Batch::Loader
       def initialize(model, association_name)
+        super()
         @model = model
         @association_name = association_name
         validate
