+++ lib/hq/graphql/enum/sort_by.rb
@@ -1,10 +1,14 @@
 # frozen_string_literal: true
 
-require "hq/graphql/enum"
+require "hq/graphql/ext/enum_extensions"
 
 module HQ
-  class GraphQL::Enum::SortBy < ::HQ::GraphQL::Enum
-    value "CreatedAt", value: :created_at
-    value "UpdatedAt", value: :updated_at
+  module GraphQL
+    module Enum
+      class SortBy < ::GraphQL::Schema::Enum
+        value "CreatedAt", value: :created_at
+        value "UpdatedAt", value: :updated_at
+      end
+    end
   end
 end
