+++ lib/hq/graphql/enum/sort_order.rb
@@ -1,10 +1,14 @@
 # frozen_string_literal: true
 
-require "hq/graphql/enum"
+require "hq/graphql/ext/enum_extensions"
 
 module HQ
-  class GraphQL::Enum::SortOrder < ::HQ::GraphQL::Enum
-    value "ASC", value: :asc
-    value "DESC", value: :desc
+  module GraphQL
+    module Enum
+      class SortOrder < ::GraphQL::Schema::Enum
+        value "ASC", value: :asc
+        value "DESC", value: :desc
+      end
+    end
   end
 end
