+++ lib/hq/graphql/ext.rb
@@ -1,0 +1,6 @@
+require "hq/graphql/ext/active_record_extensions"
+require "hq/graphql/ext/enum_extensions"
+require "hq/graphql/ext/input_object_extensions"
+require "hq/graphql/ext/mutation_extensions"
+require "hq/graphql/ext/object_extensions"
+require "hq/graphql/ext/schema_extensions"
