+++ lib/hq/graphql/root_query.rb
@@ -2,13 +2,13 @@
 
 module HQ
   module GraphQL
-    class RootQuery < ::HQ::GraphQL::Object
+    class RootQuery < ::GraphQL::Schema::Object
       def self.inherited(base)
         super
         base.class_eval do
           lazy_load do
-            ::HQ::GraphQL.root_queries.each do |field_name:, resolver:, model_name:|
-              field field_name, resolver: resolver.call, klass: model_name
+            ::HQ::GraphQL.root_queries.each do |query|
+              field query[:field_name], resolver: query[:resolver].call, klass: query[:model_name]
             end
           end
         end
