+++ lib/hq/graphql/version.rb
@@ -2,6 +2,6 @@
 
 module HQ
   module GraphQL
-    VERSION = "2.1.12"
+    VERSION = "2.2.0"
   end
 end
